Update of my prior mod to rectify problem found and added PCB mount for Anthead SF with WWG2 (#137)
* Initial Data Upload

* Create readme.md

* Update readme.md

* Update readme.md

Added Images

* Update readme.md

* Update readme.md

* Update readme.md

* Update readme.md

* Update readme.md

* Pictures Modified

* Added HF and UHF Variant

* Update readme.md

* Update readme.md

* Update Repo Location for per tool probe to current location

* Added Anthead SF WWG2 Setup and Rev 9 for Sherpa Mini

* Update Readme

* Added additional readme

* Revise readme for Anthead Sherpa Mini EBB 36 PCB Mount

Updated readme to reflect improvements in revision 9, including enhanced spring steel capture and built-in supports.

* Updated Images

* Replace images in readme with updated toolhead images

Updated images in the readme to reflect new toolhead.

* Updated Images

* Fix image filenames in readme.md

Updated image filenames in the readme for consistency.

* Update Remove Images

* Update readme for Anthead PCB mounting solution

Revised the readme to clarify the purpose and improvements of the PCB mounting solution for Anthead with Sherpa Mini and EBB36. Updated images and removed outdated revision notes.

* Update Readme

* Fix image filename in readme

Updated image filename in the readme for consistency.

* Fix model name in README for PCB Mount solution

Updated the title to reflect the correct model name.

* Fix Readme

* Revise readme for Anthead Sherpa Mini EBB 36 PCB Mount

Updated readme to reflect changes in revision 9, including fixes for torsion twist and enhancements for spring steel capturing.

* Clean up readme by removing duplicate images

Removed duplicate images from the readme file.